dashboard/requirements.txt
python-dotenv
pymssql
pandas
streamlit
altair
google-generativeai